% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbadGstats.R
\name{fbadGstats}
\alias{fbadGstats}
\title{Reports statistics across ads from Facebook Ads Manager exported data.}
\usage{
fbadGstats(filerd = "", choosedir = "NO", sumvar = "", filtervar = "",
  spentlim = 0, minevent = 0, prtrow = 20, tblout = "BOTH",
  grphout = "YES", ctrstats = "NO")
}
\arguments{
\item{filerd}{Filename to read. Must be a CSV file.}

\item{choosedir}{[Windows-Only] If 'YES', prompts the user for a directory / folder in which all CSV files will be processed by fbadGstats.}

\item{sumvar}{Variable of focus for analysis. Can provide just a few letters as long as they do not match another column in the data. For example,
could use 'REG' for 'Website Registrations Completed' (case-insensitive). Defaults to 'Link Clicks'.}

\item{filtervar}{Limits the analyzed data to only those of a primary group (the first column that appears in Ads Manager like Campaign Name or Ad Set Name)
matching the provided string of characters (case-insensitive).}

\item{spentlim}{Minimum amount that must have been spent in a breakdown group (e.g., DMA Region) in order to appear in the output.}

\item{minevent}{Minimum number of 'events' (e.g., Link Clicks) that must have occurred in order for a breakdown group to appear in the output.}

\item{prtrow}{Number of breakdown groups that will appear in the output, both the best and worst will appear separately. Default = 20.}

\item{tblout}{Show the best, worst, or both in the output? Valid values are: BEST , WORST , BOTH}

\item{grphout}{Show the bar graph and summary table of price / event (e.g., Link Clicks) for the best performers vs. all of data? Valid values are: YES or NO}

\item{ctrstats}{If yes, prints statistics for clickthrough rate if available in data (CTR) - defaults to NO. Valid values are : YES, NO. Prioritizes CTR..LINK.CLICK.THROUGH.RATE.}
}
\value{
Best or worst performing subgroups depending on `tblout` parameter. Also a graph with a complementary table for the best performers
    if grphout parameter is YES.
}
\description{
Reports statistics for breakdown groups (e.g., Age) across ads from Facebook Ads Manager exported data.
 Displays the best and worst performing breakdown groups with a ranking as well as the sum total of a specified event
    (e.g., 'Website Registrations Completed'). Can also distplay a bar graph showing the cost per specified event for the
    best breakdown groups along with a summary table comparing those best performers to the data as a whole.
}
\examples{
## to present a window in which one navigates to the desired CSV file, outputs LINK CLICKS summary by default (Performance and Clicks view suggested)
fbadGstats()
#' similar but one selects a folder and all of the CSV files are processed, and the summarized performance measure is WEBSITE.LEADS.
fbadGstats(choosedir = "YES", sumvar = "WEBSITE.LEADS")
## examine the best and worst performing Direct Marketing Areas (DMA) with respect to the results column containing 'REG' (case-insenstive, so this matches Wesbite.Registrations.Completed)
fbadGstats(filerd='example_DMA.csv', sumvar='REG')
## see more examples with output
vignette(package = "fbadstats")
}
