% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphads.R
\name{graphads}
\alias{graphads}
\title{Generates ggplot histogram and add'l statistics for fbadGstats}
\usage{
graphads(dmastat_all, prtrow, grpvar, grpvarprt, sumnam)
}
\arguments{
\item{dmastat_all}{post-processed data frame from fbadGstats}

\item{prtrow}{text}

\item{grpvar}{text}

\item{grpvarprt}{text}

\item{sumnam}{text}
}
\value{
A graph with a complementary table for the best performers
}
\description{
More text coming.
}
