% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphads.R
\name{graphads}
\alias{graphads}
\title{Generates ggplot histogram and add'l statistics for fbadGstats}
\usage{
graphads(summary_frm_distinct_grp, printrow, grpvar, grpvarprint, sumnam,
  todaydt, file_nam, spentlim, sumprintvar, sum_set)
}
\arguments{
\item{summary_frm_distinct_grp}{post-processed data frame from fbadGstats}

\item{printrow}{text}

\item{grpvar}{text}

\item{grpvarprint}{text}

\item{sumnam}{text}

\item{todaydt}{text}

\item{file_nam}{text}

\item{spentlim}{text}

\item{sumprintvar}{text}

\item{sum_set}{text}
}
\value{
A graph with a complementary table for the best performers
}
\description{
More text coming.
}
